+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * 【core   】     ListUtil.page第一页页码使用PageUtil（pr#198@Gitee）
 * 【http   】     增加微信、企业微信ua识别（pr#1179@Github）
 * 【core   】     ObjectUtil增加defaultIfXXX（pr#199@Gitee）
+* 【json   】     JSONObject构建时不支持的对象类型抛出异常
 
 ### Bug修复
 * 【core   】     修复ChineseDate没有忽略时分秒导致计算错误问题（issue#I1YW12@Gitee）
@@ -19,6 +20,7 @@
 * 【aop    】     修复SimpleAspect一个重载导致的问题，去掉重载的after方法（issue#I1YUG9@Gitee）
 * 【poi    】     修复03 sax读取日期问题（issue#I1Z83N@Gitee）
 * 【core   】     修复FileUtil.size软链导致的问题（pr#200@Gitee）
+* 【core   】     修复JSONObject构造时传入JSONArray结果出错问题（issue#I22FDS@Gitee）
 
 -------------------------------------------------------------------------------------------------------------